+++ dlshogi/utils/measure_entropy.py
@@ -28,8 +28,8 @@ BATCH_SIZE = 256
 def mini_batch(hcpevec):
     features1 = np.empty((len(hcpevec), FEATURES1_NUM, 9, 9), dtype=np.float32)
     features2 = np.empty((len(hcpevec), FEATURES2_NUM, 9, 9), dtype=np.float32)
-    move = np.empty((len(hcpevec)), dtype=np.int32)
-    result = np.empty((len(hcpevec)), dtype=np.int32)
+    move = np.empty((len(hcpevec)), dtype=np.int64)
+    result = np.empty((len(hcpevec)), dtype=np.float32)
     value = np.empty((len(hcpevec)), dtype=np.float32)
 
     cppshogi.hcpe_decode_with_value(hcpevec, features1, features2, move, result, value)
